refactor: `parseChainHubOpts` helper
